ISPN-1103 Add support for storing JSON/BSON values and metadata
Added a new 'schematic' module that contains a small document-oriented API layer on top of Infinispan and supports storing JSON and BSON documents and using JSON Schema documents to validate the structure of the documents. Rather than store the documents as strings within the Infinispan values, the value consists of a single Document with functionality that's a superset of BSON and JSON, with the goal being very efficient access and modification of the documents. Values are serialized using the BSON binary format representation. Documents are also immutable, with a simple "editable document" mechanism to allow the Infinispan values to "listen" for changes and to support DeltaAware functionality.

I had originally started to use other JSON and BSON implementations, but was not happy with the APIs. JSON libraries do not natively support binary values and other types, while BSON libraries were less mature. No JSON or BSON libraries had support for JSON Schema, though there is a relatively simplistic and mostly-complete add-on for the Jackson JSON library. Finally, implementing our own also means that we could provide an immutable API for general document access while still making it easy to create and edit documents such that DeltaAware and transaction functionality will work cleanly.

It may be useful/desirable to extract the JSON/BSON/JSON Schema functionality into a separate module or even a separate project (as it is likely useful in projects other than Infinispan).